+++ EssentialApp/EssentialAppTests/FeedImageDataLoaderWithFallbackCompositeTests.swift
@@ -6,44 +6,8 @@
 //
 
 import EssentialFeed
+import EssentialApp
 import XCTest
-
-class FeedImageDataLoaderWithFallbackComposite: FeedImageDataLoader {
-    
-    private let primary: FeedImageDataLoader
-    private let fallback: FeedImageDataLoader
-    
-    init(
-        primary: FeedImageDataLoader,
-        fallback: FeedImageDataLoader
-    ) {
-        self.primary = primary
-        self.fallback = fallback
-    }
-    
-    // MARK: - TaskWrapper
-    
-    private class TaskWrapper: FeedImageDataLoaderTask {
-        var wrapped: FeedImageDataLoaderTask?
-        
-        func cancel() {
-            wrapped?.cancel()
-        }
-    }
-    
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> any FeedImageDataLoaderTask {
-        let task = TaskWrapper()
-        task.wrapped = primary.loadImageData(from: url, completion: { [weak self] result in
-            switch result {
-            case .success:
-                completion(result)
-            case .failure:
-                task.wrapped = self?.fallback.loadImageData(from: url, completion: completion)
-            }
-        })
-        return task
-    }
-}
 
 final class FeedImageDataLoaderWithFallbackCompositeTests: XCTestCase {
